mosaic_prototype/backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
pydantic==2.7.4
pytest==8.2.1
pytest-cov==5.0.0
